+++ src/services/sections/ReturnTransaction.section.js
@@ -7,8 +7,9 @@ export default class ReturnTransactionSection {
     constructor(options) {
         this.$htmlCreator = new HtmlCreator();
         this.$translate = new TlogDocsTranslateService(options);
-        this.$utils = new Utils();
-        this._isUS = options.isUS;
+        this.$utils = new Utils(options);
+        this.realRegion = options.realRegion || 'il';
+        this.timezone = options.timezone;
     }
 
     get(options) {
@@ -26,9 +27,10 @@ export default class ReturnTransactionSection {
         elementChildren.push(elementReturnTransactionText);
 
         let originalOrderDate = this.$utils.toDate({
-            date: variables.SOURCE_ORDER_BUSINESS_DATE,
-            isUS: this._isUS,
-            format: this._isUS ? 'MM/DD/YYYY' : 'DD/MM/YYYY',
+            date: variables.CREATED_AT,
+            realRegion: this.realRegion,
+            timezone: this.timezone,
+            withoutTime: true
         });
 
         let originalOrderReferenceElementValue = this.$translate.getText(