+++ src/tlog-docs-template/billService.js
@@ -1,16 +1,15 @@
 import Utils from '../helpers/utils.service';
 import TlogDocsTranslateService from './tlogDocsTranslate';
+import Localization from "../helpers/localization.service";
 import moment from 'moment';
 import _ from "lodash";
 
 export default class BillService {
     constructor(options) {
-
-
-        this._isUS = options.isUS === undefined ? true : options.isUS;
-        this._locale = options.locale;
+        this.realRegion = options.realRegion || 'il';
         this.$utils = new Utils();
         this.$translate = new TlogDocsTranslateService(options);
+        this.$localization = new Localization(options);
     }
 
     Enums() {
@@ -69,7 +68,7 @@ export default class BillService {
         }
 
         let isTaxExempt = false;
-        if ((collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 && this._isUS)) {
+        if ((collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 && this.$localization.allowByRegions(['us', 'au']))) {
             isTaxExempt = true;
         }
 
@@ -310,9 +309,9 @@ export default class BillService {
 
     resolveTotals(variables, collections) {
         let totals = [];
-        let cashTotal = null;
+        const subtotalDiffersFromTotal = variables.TOTAL_AMOUNT !== variables.INCLUSIVE_NET_AMOUNT;
      
-        if(this._isUS) {
+        if(this.$localization.allowByRegions(['us', 'au'])) {
             let INCLUSIVE_GROSS_AMOUNT = _.get(variables, 'INCLUSIVE_GROSS_AMOUNT', variables.TOTAL_SALES_AMOUNT);
             let totalSales = _.get(variables, 'INCLUSIVE_NET_AMOUNT', variables.TOTAL_SALES_AMOUNT);
             if(INCLUSIVE_GROSS_AMOUNT !== totalSales) {
@@ -321,13 +320,7 @@ export default class BillService {
                     amount: this.$utils.toFixedSafe(INCLUSIVE_GROSS_AMOUNT, 2)
                 });
             }
-            // let DISCOUNT = _.get(variables, 'TOTAL_DISCOUNTS', 0);
-            // if (DISCOUNT > 0) {
-            //     totals.push({
-            //         name: this.$translate.getText('ORDER_DISCOUNT_US'),
-            //         amount: this.$utils.toFixedSafe(DISCOUNT * -1, 2)
-            //     })
-            // }
+
             const discounts = _.get(collections, 'ORDER_DISCOUNTS_LIST', []);
             discounts.forEach(item => {
                 totals.push({
@@ -338,9 +331,9 @@ export default class BillService {
         }
 
         if (variables.TOTAL_SALES_AMOUNT !== undefined && ((collections.ORDER_DISCOUNTS_LIST && collections.ORDER_DISCOUNTS_LIST.length > 0) ||
-            variables.TOTAL_TIPS !== undefined ||
-            (this._isUS && collections.EXCLUSIVE_TAXES && collections.EXCLUSIVE_TAXES.length > 0)) ||
-            (this._isUS && collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 ) && _.get(variables,'TOTAL_FEES', null)) {
+            variables.TOTAL_TIPS !== undefined || (this.$localization.allowByRegions(['au']) && subtotalDiffersFromTotal)
+            || (this.$localization.allowByRegions(['us']) && collections.EXCLUSIVE_TAXES && collections.EXCLUSIVE_TAXES.length > 0)) ||
+            (this.$localization.allowByRegions(['us']) && collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 ) && _.get(variables,'TOTAL_FEES', null)) {
 
             /**
              * SUBTOTAL (TOTAL_ORDER) :
@@ -349,7 +342,7 @@ export default class BillService {
              *  Backward compatibility - default is 'TOTAL_SALES_AMOUNT'
              */
             let TOTAL_SALES = 0;
-            if (this._isUS) {
+            if (this.$localization.allowByRegions(['us', 'au'])) {
                 TOTAL_SALES = _.get(variables, 'INCLUSIVE_NET_AMOUNT', variables.TOTAL_SALES_AMOUNT);
 
             } else {
@@ -361,7 +354,7 @@ export default class BillService {
                 amount: this.$utils.toFixedSafe(TOTAL_SALES, 2)
             });
         }
-        if ( !this._isUS) {
+        if (this.$localization.allowByRegions(['il'])) {
             if (collections.ORDER_DISCOUNTS_LIST && collections.ORDER_DISCOUNTS_LIST.length > 0) {
                 collections.ORDER_DISCOUNTS_LIST.forEach(discount => {
                     totals.push({
@@ -372,14 +365,24 @@ export default class BillService {
             }
         }
 
-        if (this._isUS && _.get(variables,'TOTAL_FEES', null)) {
+        if (this.$localization.allowByRegions(['us']) && _.get(variables,'TOTAL_FEES', null)) {
             totals.push({
                 name: this.$translate.getText('FEE'),
                 amount: this.$utils.toFixedSafe(variables.TOTAL_FEES , 2)
             })
         }
 
-        if (collections.EXCLUSIVE_TAXES && collections.EXCLUSIVE_TAXES.length > 0 && this._isUS) {
+        if (this.$localization.allowByRegions(['au'])) {
+            const fees = _.get(collections, 'FEES', []);
+            fees.forEach(fee => {
+                totals.push({
+                    name: fee.NAME,
+                    amount: this.$utils.toFixedSafe(fee.AMOUNT, 2)
+                });
+            });
+        }
+
+        if (collections.EXCLUSIVE_TAXES && collections.EXCLUSIVE_TAXES.length > 0 && this.$localization.allowByRegions(['us', 'au'])) {
             collections.EXCLUSIVE_TAXES.forEach(tax => {
                 totals.push({
                     type: 'exclusive_tax',
@@ -398,7 +401,7 @@ export default class BillService {
             if (autoGratuityTips && autoGratuityTips.length > 0) {
 
                 //Service charge
-                if (autoGratuityTips && autoGratuityTips.length > 0 && this._isUS) {
+                if (autoGratuityTips && autoGratuityTips.length > 0 && this.$localization.allowByRegions(['us', 'au'])) {
                     autoGratuityTips.forEach(tip => {
 
                         let _name = tip.NAME ? tip.NAME : this.$translate.getText('SERVICE_CHARGE')
@@ -436,7 +439,7 @@ export default class BillService {
             if (variables.TOTAL_TIPS_ON_PAYMENTS !== undefined && variables.TOTAL_TIPS_ON_PAYMENTS !== 0) { tipAmount = variables.TOTAL_TIPS_ON_PAYMENTS; }
             else if (variables.TOTAL_TIPS !== undefined && variables.TOTAL_TIPS !== 0) { tipAmount = variables.TOTAL_TIPS; }
 
-            if (tipAmount > 0 || !this._isUS) {
+            if (tipAmount > 0 || this.$localization.allowByRegions(['il'])) {
 
                 if (isServiceCharge === false) {
                     totals.push({
@@ -459,21 +462,21 @@ export default class BillService {
             }
         }
 
-        if (!this._isUS) {
+        if (this.$localization.allowByRegions(['il'])) {
             totals.push({
                 name: this.$translate.getText('TOTAL_INC_VAT'),
                 amount: this.$utils.toFixedSafe(_.get(variables,'TOTAL_BEFORE_TAX',variables.TOTAL_IN_VAT) || 0, 2)
             })
         }
 
-        if (this._isUS) {
+        if (this.$localization.allowByRegions(['us', 'au'])) {
             totals.push({
                 name: this.$translate.getText('TOTAL_INC_VAT'),
                 amount: this.$utils.toFixedSafe(variables.TOTAL_AMOUNT || 0, 2)
             })
         }
 
-        if(this._isUS) {
+        if(this.$localization.allowByRegions(['us'])) {
             if(variables.TOTAL_AMOUNT_AFTER_CASH_BONUS !== undefined) {
                 totals.push({
                     name: this.$translate.getText('TOTAL_AMOUNT_AFTER_CASH_BONUS'),
@@ -543,7 +546,7 @@ export default class BillService {
             ExemptedTaxData: []
         };
 
-        if (collections.INCLUSIVE_TAXES && collections.INCLUSIVE_TAXES.length > 0 && this._isUS) {
+        if (collections.INCLUSIVE_TAXES && collections.INCLUSIVE_TAXES.length > 0 && this.$localization.allowByRegions(['us', 'au'])) {
 
             taxes.InclusiveTaxes.push({
                 type: 'title',
@@ -561,7 +564,7 @@ export default class BillService {
             })
         }
 
-        if (collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 && this._isUS) {
+        if (collections.EXEMPTED_TAXES && collections.EXEMPTED_TAXES.length > 0 && this.$localization.allowByRegions(['us', 'au'])) {
 
             taxes.ExemptedTaxes.push({
                 type: 'title',
@@ -598,19 +601,19 @@ export default class BillService {
             else {
                 refund = this.$translate.getText('REFUND');
             }
-
         }
 
-        if (payment.P_TENDER_TYPE === 'creditCard' || payment.P_TENDER_TYPE === 'gidtCard') {
-            if (payment.CARD_TYPE && payment.LAST_4) {
-                paymentName = refund !== '' ? `${refund} (${payment.CARD_TYPE} ${payment.LAST_4})` : `${payment.CARD_TYPE} ${payment.LAST_4}`;
+        if (payment.P_TENDER_TYPE === 'creditCard') {
+            const card = this.$localization.allowByRegions(['au']) ?  payment.ISSUER : payment.CARD_TYPE;
+
+            if (card && payment.LAST_4) {
+                paymentName = refund !== '' ? `${refund} (${card} ${payment.LAST_4})` : `${card} ${payment.LAST_4}`;
             } else {
                 paymentName = payment.P_NAME;
             }
 
         } else {
-
-            if (this._isUS) {
+            if (this.$localization.allowByRegions(['us', 'au'])) {
                 paymentName = `${payment.P_NAME} ${refund}`;
             } else {
                 paymentName = `${refund} ${payment.P_NAME}`;
@@ -838,9 +841,7 @@ export default class BillService {
         return returnItems
     }
 
-    resolvePrintData(printData, isUS) {
-
-
+    resolvePrintData(printData, realRegion) {
         let DataBill = function (collections, variables, data, printByOrder, waiterDiners) {
             this.collections = printData.collections;
             this.variables = printData.variables;
@@ -865,7 +866,7 @@ export default class BillService {
         data.payments = this.resolvePayments(variables, collections, true);
 
         data.taxes = this.resolveTaxes(variables, collections, true);
-        data.isUS = isUS;
+        data.realRegion = realRegion;
 
         let printByOrder = this.resolvePrintByOrder(variables);
         let waiterDiners = this.resolveWaiterDiners(variables);