+++ src/tlog-docs-template/creditSlipService.js
@@ -12,7 +12,6 @@ export default class CreditSlipService {
         this.$utils = new Utils();
         this.$signatureService = new SignatureService();
         this._locale;
-        this._isUS;
         this._doc;
         this.timezone;
         this.$creditSection = new CreaditSection(options);
@@ -21,7 +20,7 @@ export default class CreditSlipService {
     }
     configure(options) {
         if (options.locale) this._locale = options.locale;
-        if (options.isUS) this._isUS = options.isUS;
+        this.realRegion = options.realRegion || 'il';
         this.timezone = options.timezone;
     }
 
@@ -89,7 +88,7 @@ export default class CreditSlipService {
             creditSlipTimeDiv.classList += " creditSlipTimeDiv";
             let providerPaymentDate = this.$utils.toDate({
                 timezone: this.timezone,
-                isUS: this._isUS,
+                realRegion: this.realRegion,
                 date: creditSlipDoc.PROVIDER_PAYMENT_DATE
             });
 
@@ -135,7 +134,7 @@ export default class CreditSlipService {
             let approvalText = this.$translate.getText('Approval');
             let CONFIRMATION_NUMBER = creditSlipDoc.CONFIRMATION_NUMBER;
             if (!CONFIRMATION_NUMBER) {
-                let possibleConfirmationNumber = creditSlipDoc.EMV?.filter(obj => obj.TYPE === 'Approval Code')[0];
+                let possibleConfirmationNumber = creditSlipDoc.EMV ? creditSlipDoc.EMV?.filter(obj => obj.TYPE === 'Approval Code')[0] : undefined;
                 CONFIRMATION_NUMBER = possibleConfirmationNumber ? possibleConfirmationNumber.DATA : '';
             }
             approvalDiv.innerHTML = "<div class='itemDiv'>" +