+++ src/tlog-docs-template/deliveryNoteTransactionService.js
@@ -10,7 +10,7 @@ import _ from "lodash";
 
 export default class DeliveryNoteTransactionDataService {
     constructor(options) {
-        this._isUS = options.isUS;
+        this.realRegion = options.realRegion || 'il';
         this.$translate = new TlogDocsTranslateService(options);
         this.$vatTemplateService = new VatTemplateService(options);
         this.$utils = new Utils();